cargo: bump rusqlite to latest

+++ src/rori/manager.rs
@@ -354,7 +354,7 @@ impl Manager {
                     info!("{} found from daemon but not in daemon, update db.", device);
                     let error_msg = format!("Failed to insert {} from database", device);
                     let result = Database::insert_new_device(&device, &String::new(), &String::new(), false).ok().expect(&*error_msg);
-                    db_devices.push((result, device.clone(), String::new(), String::new(), false));
+                    db_devices.push((result as i32, device.clone(), String::new(), String::new(), false));
                 }
             }
         }
